XGBoost zipcode hypothesis. PCA was applied for embeddings

+++ preprocessing/tree_preprocessor.py
@@ -1,9 +1,10 @@
 from sklearn.preprocessing import OneHotEncoder
 from sklearn.compose import ColumnTransformer
+import numpy as np
 
 def create_tree_preprocessor(cat_features):
     
-    cat_transformer = OneHotEncoder(handle_unknown="ignore")
+    cat_transformer = OneHotEncoder(handle_unknown="ignore", dtype=np.float32)
 
     preprocessor = ColumnTransformer(
         transformers=[
